Refactor: Header Token extrat 클래스 분리
